requirements.txt
torch==2.0.1+cpu
torchvision==0.15.2+cpu
diffusers==0.21.4
transformers==4.33.2
accelerate==0.23.0
flask==2.3.3
pillow==10.0.0
numpy==1.24.3
requests==2.31.0